+++ crates/icydb/src/db/mod.rs
@@ -1,9 +1,5 @@
-///
-/// Re-exports
-///
-pub use icydb_core::db::DbSession;
-
 pub mod query;
-pub mod response {
-    pub use icydb_core::db::response::{Response, ResponseExt, Row};
-}
+pub mod response;
+mod session;
+
+pub use session::{DbSession, SessionDeleteQuery, SessionLoadQuery};