+++ crates/test/build.rs
@@ -1,7 +1,7 @@
 use test_design as _;
 
 fn main() -> std::io::Result<()> {
-    icydb::build!("test_design::schema::Canister");
+    icydb::build!("test_design::schema::TestCanister");
 
     Ok(())
 }